# backend/openapi/openapi-chatgpt.yaml
openapi: 3.1.0
info:
  title: SuiteManager Public API for ChatGPT
  version: 1.4.8
  description: |
    API pública para Actions. Listados usan **producto_ia_v2**: precio por tarifas PG, ubicación, amenidades
    (metadata + inventario componentes), políticas mergeadas (empresa + propiedad), alias `precioBase`/`ciudad`/`direccion_corta`.
    Tras listar, llamar **obtenerDetalleAlojamiento** para ficha completa (galería, inventario_verificado, normas).

    **v1.3:** capa comercial en JSON: `contexto_turistico` (tipo_viaje, entorno, destacados, sugerencia_copy),
    `amenidades_publicas` (venta), `inventario_detallado` (strings), `descripcion_fuente` (`meta_o_columna` | `auto_template`),
    galería con `espacio` y `tipo_ia`, ubicación ampliada desde website/buildContext cuando exista.

    **v1.4:** `GET /api/alojamientos/detalle` acepta `checkin` + `checkout` (YYYY-MM-DD, aliases `fechaLlegada`/`fechaSalida`) y opcional `adultos`;
    si vienen fechas válidas, la respuesta incluye `precio_estimado` (misma función `publicWebsiteService.calculatePrice` que `/propiedad/:id/calcular-precio` + desglose checkout referencial).

    **v1.4.1:** `POST /api/reservas/cotizar` — cotización dry-run alineada a `POST /api/reservas` (validaciones, calendario, precio, desglose y política de cancelación; no persiste). Equivalente bajo `POST /api/public/reservas/cotizar` con los mismos headers que crear reserva.

...

...

...

...

...

...
servers:
  - url: https://suite-manager.onrender.com
paths:
...
  /api/reservas:
    post:
      operationId: crearReserva
      summary: Crea una reserva (requiere cabeceras de agente si AGENT_API_KEYS está configurado).
      description: |
        Equivalente funcional bajo **`POST /api/public/reservas`** (mismo cuerpo y códigos; rate limit puede diferir).
        Si no hay disponibilidad, la respuesta puede incluir **`lista_espera`** (alta en lista de espera unificada; consentimiento registrado como implícito en flujo IA).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                empresa_id: { type: string }
                alojamiento_id: { type: string }
                checkin: { type: string }
                checkout: { type: string }
                adultos: { type: integer }
                ninos: { type: integer }
                origen:
                  type: string
                  default: chatgpt
                huesped:
                  type: object
                  properties:
                    nombre: { type: string }
                    apellido: { type: string }
                    email: { type: string }
                    telefono: { type: string }
      responses:
        "201":
          description: Reserva creada
        "400":
          description: Validación (`MISSING_FIELDS`, `INVALID_EMAIL`, fechas, etc.).
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        "409":
          description: |
            No disponible (`NOT_AVAILABLE` u homólogo). Puede incluir **`lista_espera`** si se intentó registrar al huésped en lista de espera.
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean, example: false }
                  error: { type: string, example: NOT_AVAILABLE }
                  message: { type: string }
                  lista_espera:
                    $ref: "#/components/schemas/ListaEsperaEnRespuesta"
                additionalProperties: true
        "422":
          description: |
            Reglas de negocio sin reserva (`NO_DISPONIBLE`, `ALOJAMIENTO_REQUERIDO`, capacidad, etc.). En **`NO_DISPONIBLE`** puede incluir **`lista_espera`**.
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean, example: false }
                  error: { type: string, example: NO_DISPONIBLE }
                  message: { type: string }
                  lista_espera:
                    $ref: "#/components/schemas/ListaEsperaEnRespuesta"
                additionalProperties: true
  /api/public/reservas:
    post:
      operationId: crearReservaPublic
      summary: Igual que POST /api/reservas bajo /api/public (rate limit de reservas; cabeceras de agente).
      description: Mismo comportamiento y códigos que `POST /api/reservas`, incluido **`lista_espera`** en 409/422 cuando aplica.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                empresa_id: { type: string }
                alojamiento_id: { type: string }
                checkin: { type: string }
                checkout: { type: string }
                adultos: { type: integer }
                ninos: { type: integer }
                origen:
                  type: string
                  default: chatgpt
                huesped:
                  type: object
                  properties:
                    nombre: { type: string }
                    apellido: { type: string }
                    email: { type: string }
                    telefono: { type: string }
      responses:
        "201":
          description: Reserva creada
        "400":
          description: Validación
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        "409":
          description: No disponible; opcional `lista_espera`
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        "422":
          description: Sin reserva; opcional `lista_espera` en NO_DISPONIBLE
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
components:
  schemas:
    ListaEsperaEnRespuesta:
      type: object
      description: |
        Resultado del intento de alta en **lista de espera** cuando la reserva no se concreta por falta de disponibilidad (misma tabla y reconciliación que el panel).
        Si `registrado` es false, revisar `code` (p. ej. `WAITLIST_ERROR`, `NO_POSTGRES`, `WAITLIST_EMAIL_REQUIRED` en otros endpoints).
      properties:
        registrado: { type: boolean }
        duplicado:
          type: boolean
          description: Solo cuando `registrado` es true; ya existía una espera activa equivalente.
        code: { type: string, description: Cuando `registrado` es false. }
        message: { type: string }
      required: [registrado]
...